pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "graphrag-platform"
version = "0.1.0"
description = "Graph RAG Platform for SMEs - Multi-modal Knowledge Management"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0,<1.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "llama-index-core>=0.10.0",
    "llama-index-llms-ollama>=0.2.0",
    "llama-index-embeddings-ollama>=0.2.0",
    "llama-index-graph-stores-neo4j>=0.2.0",
    "llama-index-vector-stores-postgres>=0.2.0",
    "llama-index-readers-file>=0.1.0",
    "neo4j>=5.14.0",
    "psycopg2-binary>=2.9.0",
    "httpx>=0.26.0",
    "nest-asyncio>=1.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["api*", "configs*", "core*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
filterwarnings = ["ignore::DeprecationWarning"]
